requirements.txt
streamlit
Pillow
python-dotenv
requests
beautifulsoup4
anthropic
fastapi
uvicorn
Flask
pocketgroq